[master temperature_humidity] Fixes the issue with the go get.

+++ api/v1/models/weather.go
@@ -3,6 +3,6 @@ package models
 type WeatherMessage struct {
 	Success     string
 	Message     string
-	Temperature int
-	Humidity    int
+	Temperature float64
+	Humidity    float64
 }
